requirements.txt
certifi==2019.11.28
chardet==3.0.4
idna==2.9
requests==2.23.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
urllib3==1.25.8
